Split test_expand
